Update example for grid-auto-columns and grid-auto-rows properties
Note: I am aware that you can set grid-auto-rows and grid-template-rows at the same time.
However, the effects won't be seen unless the elements go outside of the grid area.
I would have added this, but the idea of elements going outside of the grid area isn't mentioned in the README until later.